[SEMI-MODULAR] Enhanced Mob Sleep (#4525)
* Add timed sleep. Differentiate voluntary sleep (sleep verb) and involuntary sleep. Allow resisting awake from timed sleep.

* Add missing into to README.md

* Tidying up overrides into module, improve documentation, add pause_expiry var

* Fix indefinite sleep bug when client was never connected

* Fully replace mob_sleep() definition

* Reduce hunger rate when lying down or sleeping

* Prolong sleep indefinitely when the client disconnects, add Click() handler to sleep status effect, change sleep status effect description

* Cleaning up

* Move files to new directory, improve readme

* Cleaning up comments

* Update readme

* Update readme

* Finish readme

* Remove erroneous proc call

* Cleanup docs

* Cleanup edit comments

* Fix bug preventing indefinite voluntary sleeping

* Update README.md

* Add type safety

* Standardize usage of STATUS_EFFECT_PERMANENT

* Refine enhanced sleep module, add signal usage

* Refine README

* Update README.md

* Update code/modules/mob/living/living.dm


* Update code/modules/mob/living/living.dm


---------

+++ code/datums/status_effects/_status_effect.dm
@@ -131,7 +131,7 @@
 		return
 
 	if(duration != STATUS_EFFECT_PERMANENT)
-		if(duration < world.time)
+		if((duration < world.time) && !pause_expiry) // NOVA EDIT: Enhanced sleep - ORIGINAL: if(duration < world.time)
 			qdel(src)
 			return
 		update_shown_duration()
@@ -213,7 +213,7 @@
 		return FALSE
 
 	duration -= seconds
-	if(duration <= world.time)
+	if((duration <= world.time) && !pause_expiry) // NOVA EDIT: Enhanced sleep - ORIGINAL: if(duration <= world.time)
 		qdel(src)
 		return TRUE
